backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bottle==0.12.25
Django==5.0.6
djangorestframework==3.15.1
gunicorn==22.0.0
lxml==5.2.2
packaging==24.0
pillow==10.3.0
proxy_tools==0.1.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
